전체 검색, 인기검색어 관리/순위 개발 완료

+++ app/Http/Controllers/Admin/PopupsController.php
@@ -21,6 +21,10 @@ class PopupsController extends Controller
      */
     public function index()
     {
+        if(auth()->user()->cant('index', $this->popup)) {
+            abort(403, '팝업 레이어 관리 목록 보기에 대한 권한이 없습니다.');
+        }
+
         $params = $this->popup->getIndexParams();
 
         return view('admin.popups.index', $params);
@@ -33,6 +37,10 @@ class PopupsController extends Controller
      */
     public function create()
     {
+        if(auth()->user()->cant('create', Popup::class)) {
+            abort(403, '팝업 레이어 생성에 대한 권한이 없습니다.');
+        }
+
         $params = $this->popup->getCreateParams();
 
         return view('admin.popups.form', $params);
@@ -46,6 +54,10 @@ class PopupsController extends Controller
      */
     public function store(Request $request)
     {
+        if(auth()->user()->cant('create', Popup::class)) {
+            abort(403, '팝업 레이어 생성에 대한 권한이 없습니다.');
+        }
+
         $result = $this->popup->storePopup($request);
 
         return redirect(route('admin.popups.edit', $result->id));
@@ -59,6 +71,10 @@ class PopupsController extends Controller
      */
     public function edit($id)
     {
+        if(auth()->user()->cant('update', $this->popup)) {
+            abort(403, '팝업 레이어 수정에 대한 권한이 없습니다.');
+        }
+
         $params = $this->popup->getEditParams($id);
 
         return view('admin.popups.form', $params);
@@ -73,6 +89,10 @@ class PopupsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(auth()->user()->cant('update', $this->popup)) {
+            abort(403, '팝업 레이어 수정에 대한 권한이 없습니다.');
+        }
+
         $result = $this->popup->updatePopup($request, $id);
 
         return redirect(route('admin.popups.edit', $id));
@@ -86,6 +106,10 @@ class PopupsController extends Controller
      */
     public function destroy($id)
     {
+        if(auth()->user()->cant('delete', $this->popup)) {
+            abort(403, '팝업 레이어 삭제에 대한 권한이 없습니다.');
+        }
+
         $result = $this->popup->deletePopup($id);
 
         return redirect()->back();
